Update README.md
Archiving project, since JOVO has been archived. It should be possible to revive this by interfacing directly with the Alexa libraries, retaining the database back-end... but I'm not sure when that will happen.

WNYC recently threatened to kill the New Sounds program. It was rescued the same way WNYC itself was when it lost NYC support, by an outpouring of financial support from listeners that purchased a few years of breathing room for it to seek other funding. This budgeting shift may affect the web accesses I've been relying upon to browse and play the New Sounds archives and livestream. Only time will tell. 

The question is, is there enough interest in New Sounds, and in accessing it via Alexa, to justify migrating this Skill?